Removing faulty import
As pointed out in #14, the import in this is breaking pip because it tries to import requests before it's installed. As far as I can see, this import doesn't do anything, and it installed fine without it. So I request that this line gets removed so this package gets installed correctly through pip.